refactor: Split EntityWriteBatch into per-entity writers [DHIS2-21378]
