+++ dhis-2/dhis-tracker/src/main/java/org/hisp/dhis/tracker/imports/bundle/persister/RelationshipPersister.java
@@ -69,6 +69,23 @@ public class RelationshipPersister
   }
 
   @Override
+  protected void assignId(org.hisp.dhis.tracker.model.Relationship entity, long id) {
+    entity.setId(id);
+  }
+
+  @Override
+  protected void stageInsert(
+      org.hisp.dhis.tracker.model.Relationship entity, EntityWriteBatch batch) {
+    batch.stageInsert(entity);
+  }
+
+  @Override
+  protected void stageUpdate(
+      org.hisp.dhis.tracker.model.Relationship entity, EntityWriteBatch batch) {
+    throw new UnsupportedOperationException("Relationships are not updated");
+  }
+
+  @Override
   protected org.hisp.dhis.tracker.model.Relationship convert(
       TrackerBundle bundle, Relationship trackerDto) {
     if (bundle.getStrategy(trackerDto) == TrackerImportStrategy.UPDATE) {
