+++ dhis-2/dhis-tracker/src/main/java/org/hisp/dhis/tracker/imports/bundle/persister/UpsertTableWriter.java
@@ -1,0 +1,85 @@
+/*
+ * Copyright (c) 2004-2026, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its contributors 
+ * may be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.tracker.imports.bundle.persister;
+
+import static org.hisp.dhis.tracker.imports.bundle.persister.JdbcBatchSupport.truncate;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Base for the per-entity writers that stage inserts and updates of a single top-level entity type
+ * (TrackedEntity, Enrollment, TrackerEvent, SingleEvent). Owns the two staging lists and the
+ * mark/rollback/clear lifecycle shared with {@link ChangeLogAccumulator}; subclasses supply the
+ * type-specific {@link #flush(Connection)} (multi-row INSERT, unnest UPDATE and any notes cascade).
+ */
+abstract class UpsertTableWriter<E> {
+
+  protected final List<E> inserts = new ArrayList<>();
+  protected final List<E> updates = new ArrayList<>();
+
+  void stageInsert(E entity) {
+    inserts.add(entity);
+  }
+
+  void stageUpdate(E entity) {
+    updates.add(entity);
+  }
+
+  /** Whether the given entity is staged for insert (no DB row yet) in this batch. */
+  boolean isStagedAsInsert(E entity) {
+    return inserts.contains(entity);
+  }
+
+  Mark mark() {
+    return new Mark(inserts.size(), updates.size());
+  }
+
+  void rollbackTo(Mark mark) {
+    truncate(inserts, mark.inserts());
+    truncate(updates, mark.updates());
+  }
+
+  void clear() {
+    inserts.clear();
+    updates.clear();
+  }
+
+  boolean isEmpty() {
+    return inserts.isEmpty() && updates.isEmpty();
+  }
+
+  /** Applies the staged inserts and updates via JDBC on {@code conn}. */
+  abstract void flush(Connection conn) throws SQLException;
+
+  record Mark(int inserts, int updates) {}
+}
